SD Card via MSC working.
Trying to organize to be able to use the block device for filesystem access when the USB is not connected... has proven troublesome. The macros to create the interfaces/devices/device classes are C-only because of some header jazz, and seem to only like to live in global scope. I might be able to scapegoat them into a function later, but this "global" folder will do for now. Next up is reading/writing files to/from the SD card when I'm not in USB mode. Also to be able to react differently on if I'm in USB mode or not :)