Migrated ILI9341 to nrfx_spim, moved to 32 MHz SPI.
Almost doubled the speed of the draw test. I think the limiting factor now is the CPU, bus is probably quick enough for the most part. Could optimize it further by removing all the single byte buffer calls and replacing with larger buffers, but I'll profile it later if it really matters.