Menus are getting functional!
Renamed `main_menu` to `list_menu` as it should be usable for any list selection screen. Boot goes to a MenuMain state which has a few hardcoded options.

Up next: The transition isn't there yet, but the 'Load Payload' option should hook to the MenuFiles state, which will allow directory browsing. Then to get the MVP working, pressing on a file should set it as the payload, activate the NFCT, and switch to that state! Leaving that state will deactivate the NFCT, and drop back to the file browsing, which I should hopefully be able to save my state.