Directory navigation working!
I thought I'd cheat and use a std::deque and std::strings to manage the path instead of twiddling the strings myself. Well, it worked, but it basically doubled the size of my flash use, so I'm ripping that out now.

Also, it crashes on directories with more than 64 files at the moment, because I haven't done pagination yet.