Add bootloader.
And work it into the main makefile. Integration isn't perfect for copying the mbr'd bootloader hex out, but it works on Windows for now.

Flashing will now flash the bootloader merged with the application, and then use pyocd to poke 'app valid' and a crc of 0x0000 into the bootloader settings page (0xFF000). It's a little 'hardcodeish', but it's the way adafruit does it for circuit python for now, so it'll do.