Littlefs working on flash!
Lots, lots of changes to make it happen. I hate doing this in one big commit, but it's sorta an all or nothing.

Fleshed out the settings to a few more things that will make sense. I don't love doing the raw structure read/write, but My other solution is an INI file and base64ing the binary bits, and that's just too much effort for now. Did find a decent library or two, we shall see. I'll probably still end up going with a MsgPack library of sorts eventually, unless there's a nice c json that does binary out of the box, so I don't have to do base64 roundtrips.