Add sleep command for ILI9341 on SYSOFF.
Based on my (not terribly accurate) power supply, it looks like I'm drawing ~110-160 mA in normal operation, and now a teeny 2 mA in SYSOFF. It was 11 mA before the LCD sleep mode, so I imagine that if I had a way of disconnecting it completely it'd be even lower. I'm going to investigate if the Wi-Fi chip is pulling any despite being unused, but I'm pretty happy with 2 mA. With a 500 mAh battery, that's a theoretical 10+ days of standby time, which I think I can live with for now!