Converted low power pwm to a rgb led driver.
Would need a lot bit of work to use with actual RGB LEDs, but should work with a the backlight.

Also changed lots of effects/led related stuff. There's a lot going on here, but mostly it's to try and share the architecture for effects and drivers between RGB Leds and single colors.

Still going to have to move to real PWM for the backlight I think. The low power (even without Cie1931) still flickers.